requirements.txt
altair>=5.3.0
gradio>=4.27.0
lightgbm>=4.3.0
matplotlib>=3.8.2
numpy>=1.26.4
pandas>=2.2.2
Pillow>=10.3.0
pytorch_lightning>=2.1.3
Requests>=2.31.0
timm>=0.9.16
torch>=2.2.2
torchinfo>=1.8.0
torchmetrics>=1.2.1
torchvision>=0.17.2
tqdm>=4.66.1
tensorboard>=2.15.1
